Add function ZenUnit::ZenUnitTestRunner::SetNextGlobalZenUnitModeRandomSeed() for updating the random seed for the next test run when --test-runs=N is specified

+++ ZenUnitUtilsAndAssertionTests/Components/Time/MetalMock/WatchMock.h
@@ -5,5 +5,5 @@ class WatchMock : public Metal::Mock<ZenUnit::Watch>
 public:
    METALMOCK_NONVOID0_CONST(string, DateTimeNow)
    METALMOCK_NONVOID0_CONST(tm, TMNow)
-   METALMOCK_NONVOID0_CONST(long long, SecondsSince1970)
+   METALMOCK_NONVOID0_CONST(unsigned, SecondsSince1970)
 };
